Safari-609.1.11.1/LayoutTests/fast/css/parsing-paint-order-expected.txt

This tests checks that all of the input values for paint-order parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyleInherited("paint-order: fill;") is "fill"
PASS testComputedStyleInherited("paint-order: stroke;") is "stroke"
PASS testComputedStyleInherited("paint-order: markers;") is "markers"
PASS testComputedStyle(";") is "normal"
PASS test("paint-order: fill;") is "fill"
PASS test("paint-order: fill stroke;") is "fill"
PASS test("paint-order: fill markers;") is "fill markers"
PASS test("paint-order: fill stroke markers;") is "fill"
PASS test("paint-order: fill markers stroke;") is "fill markers"
PASS test("paint-order: stroke;") is "stroke"
PASS test("paint-order: stroke fill;") is "stroke"
PASS test("paint-order: stroke markers;") is "stroke markers"
PASS test("paint-order: stroke fill markers;") is "stroke"
PASS test("paint-order: stroke markers fill;") is "stroke markers"
PASS test("paint-order: markers;") is "markers"
PASS test("paint-order: markers fill;") is "markers"
PASS test("paint-order: markers stroke;") is "markers stroke"
PASS test("paint-order: markers fill stroke;") is "markers"
PASS test("paint-order: markers stroke fill;") is "markers stroke"
PASS test("paint-order: fil;") is ""
PASS test("paint-order: fill markrs;") is ""
PASS test("paint-order: 10px;") is ""
PASS test("paint-order: 10%;") is ""
PASS successfullyParsed is true

TEST COMPLETE
